masks/previous/requirements_mask.txt
annexremote==1.4.3
appdirs==1.4.4
bids-validator==1.3.12
boto==2.49.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
cryptography==3.2.1
datalad==0.13.5
decorator==4.4.1
Deprecated==1.2.10
docopt==0.6.2
etelemetry==0.1.2
fasteners==0.15
filelock==3.0.12
future==0.18.2
humanize==2.6.0
idna==2.8
importlib-metadata==2.0.0
iso8601==0.1.13
isodate==0.6.0
jeepney==0.4.3
joblib==0.14.1
keyring==20.0.1
keyrings.alt==3.2.0
lxml==4.4.2
monotonic==1.5
msgpack==1.0.0
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.0
niflow-nipype1-workflows==0.0.4
nilearn==0.6.0
nipype==1.4.0
num2words==0.5.10
numpy==1.18.0
packaging==20.0
pandas==0.25.3
patool==1.12
patsy==0.5.1
prov==1.5.3
pybids==0.10.0
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
PyGithub==1.53
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
rdflib==4.2.2
requests==2.22.0
scikit-learn==0.22.1
SecretStorage==3.2.0
simplejson==3.17.0
six==1.13.0
sklearn
SQLAlchemy==1.3.12
tqdm==4.51.0
traits==5.2.0
urllib3==1.25.7
Whoosh==2.7.4
wrapt==1.12.1
zipp==1.2.0
